requirements.txt
requests
python-dotenv
flask
boto3
confluent-kafka
kafka-python